updated index.ts of multieditor
